Preliminary Java Objects
Very shakey versions of the objects we'll be using. I see what you mean
about not really needing specific objects in a WebApp now, but I think
we should hold on to them unless we manage to get an implementation
going that compeltely excludes anything except queries and their
results.

These are 'just-in-case' classes for now.